refactor: Separate PrometheusHandler registration to its own route
